Create noenter.cpp
We create a simulated login that will prompt
password and issue a prompt only if the entered password matches the one predetermined in the program, and in case of an error, issue a message asking you to repeat the entry. After three errors, the program should display a message about an unauthorized access attempt.