Create tables_of_values.cpp
display a table of values of the function Y(x) and its
series expansion S(x) for x varying from xn to xk in steps
h = (xk – xn)/10. The closeness of the values of S(x) and Y(x) over the entire range of x values indicates the correctness of the calculation of S(x) and Y(x).